Made a util filtering and restricting venues displaying if criterias not met, so that both VenuesPage and HomePage can use it. Also made some small styling changes, ex background-image homepage, some font-weights and div positioning
