+++ src/pages/venues-page/VenuesPage.jsx
@@ -2,67 +2,44 @@ import React, { useEffect, useState } from "react";
 import * as S from "./VenuesPage.styles";
 import VenuesCard from "../../components/venues-card/VenuesCard";
 import useApi from "../../hooks/useApi";
-import { shuffle, handleSortChange, sortVenues } from "../../utils/venues-utils/venuesUtils";
+import { shuffle, handleSortChange, sortVenues, filterValidVenues } from "../../utils/venues-utils/venuesUtils";
 
-const fetchAllVenues = async (fetchApi) => {
-  let allVenues = [];
-  let page = 1;
-  const limit = 100;
-  let totalFetched = 0;
-
-  do {
-    try {
-      const result = await fetchApi(`https://v2.api.noroff.dev/holidaze/venues?limit=${limit}&page=${page}`, "GET");
-      if (result.data && Array.isArray(result.data)) {
-        totalFetched = result.data.length;
-        allVenues = [...allVenues, ...result.data];
-        page++;
-      } else {
-        totalFetched = 0;
-      }
-    } catch (error) {
-      console.error("Fetching venues failed:", error);
-      break;
-    }
-  } while (totalFetched > 0);
-
-  return allVenues;
-};
-
-const VenuesPage = () => {
+function VenuesPage() {
   const { fetchApi, isLoading, isError } = useApi();
   const [venues, setVenues] = useState([]);
   const [sort, setSort] = useState("");
 
   useEffect(() => {
-    fetchAllVenues(fetchApi)
-      .then((venues) => {
-        const validVenues = venues.filter((venue) => {
-          const hasValidMedia = venue.media && venue.media.length > 0 && !venue.media[0].url.includes("string");
-          const hasValidPrice = venue.price != null && !venue.price.toString().includes("string");
-          const hasValidName = venue.name && !venue.name.includes("string");
-          const hasValidLocation = venue.location && venue.location.city && !venue.location.city.includes("string");
-          const hasValidCountry = venue.location && (venue.location.country === "Norway" || venue.location.country === "Norge");
+    const fetchAllVenues = async () => {
+      let allVenues = [];
+      let page = 1;
+      const limit = 100;
+      let totalFetched = 0;
 
-          return hasValidMedia && hasValidPrice && hasValidName && hasValidLocation && hasValidCountry;
-        });
-
-        const uniqueImageUrls = new Set();
-        const uniqueVenues = validVenues.filter((venue) => {
-          const imageUrl = venue.media[0].url;
-          if (!uniqueImageUrls.has(imageUrl) && !imageUrl.includes("string")) {
-            uniqueImageUrls.add(imageUrl);
-            return true;
+      do {
+        try {
+          const result = await fetchApi(`https://v2.api.noroff.dev/holidaze/venues?limit=${limit}&page=${page}`, "GET");
+          if (result.data && Array.isArray(result.data)) {
+            totalFetched = result.data.length;
+            allVenues = [...allVenues, ...result.data];
+            page++;
+          } else {
+            totalFetched = 0;
           }
-          return false;
-        });
+        } catch (error) {
+          console.error("Fetching venues failed:", error);
+          break;
+        }
+      } while (totalFetched > 0);
 
-        const shuffledVenues = shuffle(uniqueVenues);
-        setVenues(shuffledVenues);
-      })
-      .catch((error) => {
-        console.error("Fetching all venues failed:", error);
-      });
+      return allVenues;
+    };
+
+    fetchAllVenues().then((venues) => {
+      const validVenues = filterValidVenues(venues);
+      const shuffledVenues = shuffle(validVenues);
+      setVenues(shuffledVenues);
+    });
   }, [fetchApi]);
 
   const sortedVenues = venues.length > 0 ? sortVenues(venues, sort) : [];
@@ -99,6 +76,6 @@ const VenuesPage = () => {
       </S.VenuesGrid>
     </S.PageContainer>
   );
-};
+}
 
 export default VenuesPage;
